requirements.txt
Flask
gunicorn
pytesseract
PyMuPDF
Pillow
Werkzeug